Move sku variant related functions to a separate file
